Added statistics update, fixed sharedpref getlong for sync date, creation default date

+++ app/src/main/java/br/ufrgs/inf01059/wikipapers/PaperListActivity.java
@@ -87,7 +87,7 @@ public class PaperListActivity extends ActionBarActivity implements
 		if (sharedPref.contains("nSyncNotes") == false) {
 			SharedPreferences.Editor editor = sharedPref.edit();
 			editor.putInt("nSyncNotes", 0);
-			editor.putInt("syncDate", 0);
+			editor.putLong("syncDate", 0);
 			editor.commit();
 		}
